utf8: fix pointer type mismatch

+++ src/utf8.c
@@ -75,7 +75,7 @@ int
 ncd_utf8_decode(Codepoint *const restrict cp, char const *const restrict u8str,
     size_t const n)
 {
-	uint_least8_t processed;
+	size_t processed;
 	uint_least8_t const len = ncd_utf8_unit_count(u8str);
 
 	if (!cp)
